Statisticが追加されたときの描画バグを修正。

+++ src/aoetk/fxglassfishmonitor/model/MetricType.java
@@ -5,5 +5,5 @@ package aoetk.fxglassfishmonitor.model;
  * @author aoetk
  */
 public enum MetricType {
-    LONG, DATETIME, STRING
+    INTEGER, DATETIME, STRING
 }
